Move testing/output to public/node
