+++ packages/cli-hydrogen/src/cli/commands/hydrogen/info.test.ts
@@ -1,7 +1,7 @@
 import InfoCommand from './info.js'
 import {HydrogenApp, load as loadApp} from '../../models/hydrogen.js'
 import {describe, expect, vi, it, beforeAll} from 'vitest'
-import {outputMocker} from '@shopify/cli-kit'
+import {mockAndCaptureOutput} from '@shopify/cli-kit/node/testing/output'
 
 beforeAll(() => {
   vi.mock('../../models/hydrogen')
@@ -42,7 +42,7 @@ function mockOutput(mockHydrogenApp: Partial<HydrogenApp> = {}) {
 
   vi.mocked(loadApp).mockResolvedValue(app)
 
-  const mocker = outputMocker.mockAndCaptureOutput()
+  const mocker = mockAndCaptureOutput()
 
   return {
     ...mocker,
