+++ packages/cli-kit/src/index.ts
@@ -3,4 +3,3 @@ export * as output from './output.js'
 export * as schema from './schema.js'
 export * as store from './store.js'
 export * as ui from './ui.js'
-export * as outputMocker from './testing/output.js'
