+++ packages/cli/src/cli/commands/auth/logout.test.ts
@@ -1,7 +1,7 @@
 import Logout from './logout.js'
 import {beforeEach, describe, expect, it, vi} from 'vitest'
-import {outputMocker} from '@shopify/cli-kit'
 import {logout} from '@shopify/cli-kit/node/session'
+import {mockAndCaptureOutput} from '@shopify/cli-kit/node/testing/output'
 
 beforeEach(() => {
   vi.mock('@shopify/cli-kit/node/session')
@@ -10,7 +10,7 @@ beforeEach(() => {
 describe('logs out', () => {
   it('clears the session', async () => {
     // Given
-    const outputMock = outputMocker.mockAndCaptureOutput()
+    const outputMock = mockAndCaptureOutput()
 
     // When
     await Logout.run()
